kis_api_backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
pytest
httpx
pytest-httpx

# Database
sqlmodel==0.0.31
alembic==1.18.1
google-cloud-firestore==2.18.0

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
email-validator==2.3.0

# Encryption
cryptography==44.0.0
